+++ tests/test_code/issue140/callee.py
@@ -1,0 +1,5 @@
+def dostuff():
+    pass
+
+
+dostuff()