+++ tests/test_code/issue140/caller.py
@@ -1,0 +1,3 @@
+import callee
+
+callee.dostuff()