+++ tests/test_code/issue140/construct.py
@@ -1,0 +1,9 @@
+class Thing:
+    def method(self):
+        pass
+
+
+def f():
+    obj = Thing()
+    obj.method()
+    return obj